Revert "Update sampling"
This reverts commit 88f5867fa76aaf64c80ead891913a0642085fd10.

+++ include/caffe/sampling.hpp
@@ -31,11 +31,11 @@
 
 namespace caffe {
 
-void sample_blob(const fp16* blob, int blob_count, std::map<int, int> &exp_map, std::map<int, int> &frac_map, std::map<int, int> &val_map,  std::vector<float> &val_vector, int sampling_frequency);
+void sample_blob(const fp16* blob, int blob_count, std::map<int, int> &exp_map, std::map<int, int> &frac_map, std::map<int, int> &val_map,  std::vector<uint32_t> &val_vector, int sampling_frequency);
 
-void sample_blob(const double* blob, int blob_count, std::map<int, int> &exp_map, std::map<int, int> &frac_map, std::map<int, int> &val_map,  std::vector<float> &val_vector, int sampling_frequency);
+void sample_blob(const double* blob, int blob_count, std::map<int, int> &exp_map, std::map<int, int> &frac_map, std::map<int, int> &val_map,  std::vector<uint32_t> &val_vector, int sampling_frequency);
 
-void sample_blob(const fp16* blob, int blob_count, std::map<int, int> &exp_map, std::map<int, int> &frac_map, std::map<int, int> &val_map,  std::vector<float> &val_vector, int sampling_frequency);
+void sample_blob(const fp16* blob, int blob_count, std::map<int, int> &exp_map, std::map<int, int> &frac_map, std::map<int, int> &val_map,  std::vector<uint32_t> &val_vector, int sampling_frequency);
 
 void print_map(std::map<int, int> sample_map);
 }
